Add standings data test fixture
Scrapping standings-by-date data from basketball reference is moved to a
pytest fixture so that it will occur once for all tests. This limits the
number of HTTP requests to the basketball reference website to once per
test execution per tox test environment to be kind to the basketball
reference servers.