Adjust form of processed standings data
The form of the DataFrame returned from the processed standings data was
changed to retain more of the standings data (wins, losses, games
played, etc.) instead of just returning the win fraction by date for
each team.

Necessary changes to plot_standings_progression were made as a resul of
this change.